Update .NET version in Nuspec to lib\net46
Tried to install this package to a .NET 4.5 project. It caused a very confusing error, which I tracked down to the fact that the project could not be built using the referenced Stripe.net assembly. I believe this change is what is needed to let NuGet know that it can't install this version to  4.5 projects.