Experiment with Markdown changelog
Here we experiment a bit with our changelog format by changing
`CHANGELOG` to `CHANGELOG.md` and refactoring.

The new format is largely based on keepachangelog.com [1], but I've
streamlined it a little bit (by removing subsubheadings) because I find
the default a little tedious to keep up-to-date and visually noisy
enough that it's hard to skim.

Advantages of the new format:

* Linkable headers
* Nice in-browser rendering (standard Markdown)
* Easy to copy and paste additions right into the releases page when
  cutting a new release
* Real links to all PRs that introduced changes

[1] http://keepachangelog.com/en/0.3.0/