Make "exchange rate" singular
We've recently tweaked the very new exchange rates API somewhat so that
a single "exchange rate" resource becomes singular instead of plural.
This is so that it falls into convention with all the other resources in
the API.

I jumped the gun a little bit of merging the .NET pull request, so this
one goes back through and amends a few places that were using a plural
exchange rate so that they're now singular.

Apparently we got a little lucky because 11.7.0 (the version I sent this
out under) apparently never made it to Nuget (possibly a build failure
which I'll have to look into separately). But anyway, it's safe to
release this without a breaking change because there is no official
package release with the other API.