Move to the latest API version and add new properties/params.
* Add support for tax_info and tax_info_verification on the customer
* Remove obsolete subscription service
* Remove business_vat_id on the Customer
* Rename amount to unit_amount on Plan's tiers
* Remove the ability to call Cancel on a Subscription with at_period_end.
This involved some changes to the prototypes. For that reason I added a new
options class as we will soon add parameters on subscription cancelation and
it avoids the second major change. I also took this as an opportunity to remove
the prototype with plan and customer. It did not make sense to keep supporting
this and to remove taking customer as a method parameter. The options classes
have CustomerId for that purpose.
* percent_off on Coupons is now a decimal
* Products do not that the list of SKUs embedded anymore.